--- .vite/deps_temp_8157e46e/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
