+++ src/app/[lng]/123/page.tsx
@@ -1,0 +1,9 @@
+import Image from "next/image";
+
+export default function Page() {
+  return (
+    <div>
+      <Image src="/vtal.png" alt="VTAL" width={100} height={100} />
+    </div>
+  );
+}
