+++ src/app/[lng]/blog/[slug]/page.tsx
