+++ src/app/[lng]/layout.tsx
@@ -1,27 +1,86 @@
+import Navbar from "@/components/navbar";
+import { ThemeProvider } from "@/components/theme-provider";
+import { TooltipProvider } from "@/components/ui/tooltip";
+import { DATA_en } from "@/data/resume_en";
+import { cn } from "@/lib/utils";
+import type { Metadata } from "next";
+import { Inter as FontSans } from "next/font/google";
 import { dir } from 'i18next'
 import { languages } from '../i18n/settings'
+
+import "../globals.css";
+
+const fontSans = FontSans({
+  subsets: ["latin"],
+  variable: "--font-sans",
+});
 
 export async function generateStaticParams() {
   return languages.map((lng) => ({ lng }))
 }
+
+export const metadata: Metadata = {
+  metadataBase: new URL(DATA_en.url),
+  title: {
+    default: DATA_en.name,
+    template: `%s | ${DATA_en.name}`,
+  },
+  description: DATA_en.description,
+  openGraph: {
+    title: `${DATA_en.name}`,
+    description: DATA_en.description,
+    url: DATA_en.url,
+    siteName: `${DATA_en.name}`,
+    locale: "en_US",
+    type: "website",
+  },
+  robots: {
+    index: true,
+    follow: true,
+    googleBot: {
+      index: true,
+      follow: true,
+      "max-video-preview": -1,
+      "max-image-preview": "large",
+      "max-snippet": -1,
+    },
+  },
+  twitter: {
+    title: `${DATA_en.name}`,
+    card: "summary_large_image",
+  },
+  verification: {
+    google: "",
+    yandex: "",
+  },
+};
 
 export default function RootLayout({
   children,
   params: {
     lng
   }
-}: {
-  children: React.ReactNode,
+}: Readonly<{
+  children: React.ReactNode;
   params: {
-    lng: string
-  }
-}) {
+    lng: string;
+  };
+}>) {
   return (
-    <html lang={lng} dir={dir(lng)}>
-      <head />
-      <body>
-        {children}
+    <html lang={lng} dir={dir(lng)} suppressHydrationWarning>
+      <body
+        className={cn(
+          "min-h-screen bg-background font-sans antialiased max-w-2xl mx-auto py-12 sm:py-24 px-6",
+          fontSans.variable
+        )}
+      >
+        <ThemeProvider attribute="class" defaultTheme="light">
+          <TooltipProvider delayDuration={0}>
+            {children}
+            <Navbar lng={lng} />
+          </TooltipProvider>
+        </ThemeProvider>
       </body>
     </html>
-  )
+  );
 }
