+++ src/app/[lng]/page.tsx
@@ -1,16 +1,219 @@
-import Link from 'next/link'
-import { useTranslation } from '../i18n'
-import { Footer } from './components/Footer'
+import BlurFade from "@/components/magicui/blur-fade";
+import BlurFadeText from "@/components/magicui/blur-fade-text";
+import { ProjectCard } from "@/components/project-card";
+import { ResumeCard } from "@/components/resume-card";
+import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
+import { Badge } from "@/components/ui/badge";
+import Link from "next/link";
+import Markdown from "react-markdown";
+import { useTranslation } from '../i18n';
+import { LinkIcon } from "lucide-react";
+import Image from "next/image";
+
+const BLUR_FADE_DELAY = 0.04;
 
 export default async function Page({ params: { lng } }: { params: { lng: any } }) {
-  const { t } = await useTranslation(lng, 'translation')
+  const { t } = await useTranslation(lng, 'translation');
+
+  const workExperiences = t('work.experiences', { returnObjects: true }) as Array<{
+    company: string;
+    logoUrl: string;
+    title: string;
+    href: string;
+    badges: string[];
+    start: string;
+    end?: string;
+    description: string;
+  }>;
+
+  const educationList = t('education.schools', { returnObjects: true }) as Array<{
+    school: string;
+    href: string;
+    logoUrl: string;
+    degree: string;
+    start: string;
+    end: string;
+  }>;
+
+  const skillsList = t('skills.list', { returnObjects: true }) as string[];
+
+  const projectsList = t('projects.list', { returnObjects: true }) as Array<{
+    title: string;
+    href: string;
+    description: string;
+    dates: string;
+    technologies: string[];
+    image: string;
+    links: Record<string, string>;
+  }>;
+
   return (
-    <>
-      <h1>{t('title')}</h1>
-      <Link href={`/${lng}/second-page`}>
-        {t('to-second-page')}
-      </Link>
-      <Footer lng={lng}/>
-    </>
-  )
+    <main className="flex flex-col min-h-[100dvh] space-y-10">
+      <section id="hero">
+        <div className="mx-auto w-full max-w-2xl space-y-8">
+          <div className="gap-2 flex justify-between">
+            <div className="flex-col flex flex-1 space-y-1.5">
+              <BlurFadeText
+                delay={BLUR_FADE_DELAY}
+                className="text-3xl font-bold tracking-tighter sm:text-5xl xl:text-6xl/none"
+                yOffset={8}
+                text={t('hero.greeting')}
+              />
+              <BlurFadeText
+                className="max-w-[600px] md:text-xl"
+                delay={BLUR_FADE_DELAY}
+                text={t('description')}
+              />
+            </div>
+            <BlurFade delay={BLUR_FADE_DELAY}>
+              <Avatar className="size-28 border">
+                <AvatarImage alt={t('name')} src={t('avatarUrl')} />
+                <AvatarFallback>{t('initials')}</AvatarFallback>
+              </Avatar>
+            </BlurFade>
+          </div>
+        </div>
+      </section>
+      <section id="about">
+        <BlurFade delay={BLUR_FADE_DELAY * 3}>
+          <h2 className="text-xl font-bold">About</h2>
+        </BlurFade>
+        <BlurFade delay={BLUR_FADE_DELAY * 4}>
+          <Markdown className="prose max-w-full text-pretty font-sans text-sm text-muted-foreground dark:prose-invert">
+            {t('summary')}
+          </Markdown>
+        </BlurFade>
+      </section>
+      <section id="work">
+        <div className="flex min-h-0 flex-col gap-y-3">
+          <BlurFade delay={BLUR_FADE_DELAY * 5}>
+            <h2 className="text-xl font-bold">{t('work.title')}</h2>
+          </BlurFade>
+          {workExperiences.map((work, id) => (
+            
+            <BlurFade
+              key={work.company}
+              delay={BLUR_FADE_DELAY * 6 + id * 0.05}
+            >
+              <ResumeCard
+                key={work.company}
+                logoUrl={work.logoUrl}
+                altText={work.company}
+                title={work.company}
+                subtitle={work.title}
+                href={work.href}
+                badges={work.badges}
+                period={`${work.start} - ${work.end ?? 'Present'}`}
+                description={work.description}
+              />
+            </BlurFade>
+          ))}
+        </div>
+      </section>
+      <section id="education">
+        <div className="flex min-h-0 flex-col gap-y-3">
+          <BlurFade delay={BLUR_FADE_DELAY * 7}>
+            <h2 className="text-xl font-bold">{t('education.title')}</h2>
+          </BlurFade>
+          {educationList.map((education, id) => (
+            <BlurFade
+              key={education.school}
+              delay={BLUR_FADE_DELAY * 8 + id * 0.05}
+            >
+              <ResumeCard
+                key={education.school}
+                href={education.href}
+                logoUrl={education.logoUrl}
+                altText={education.school}
+                title={education.school}
+                subtitle={education.degree}
+                period={`${education.start} - ${education.end}`}
+              />
+            </BlurFade>
+          ))}
+        </div>
+      </section>
+      <section id="skills">
+        <div className="flex min-h-0 flex-col gap-y-3">
+          <BlurFade delay={BLUR_FADE_DELAY * 9}>
+            <h2 className="text-xl font-bold">{t('skills.title')}</h2>
+          </BlurFade>
+          <div className="flex flex-wrap gap-1">
+            {skillsList.map((skill, id) => (
+              <BlurFade key={skill} delay={BLUR_FADE_DELAY * 10 + id * 0.05}>
+                <Badge key={skill}>{skill}</Badge>
+              </BlurFade>
+            ))}
+          </div>
+        </div>
+      </section>
+      <section id="projects">
+        <div className="space-y-12 w-full py-12">
+          <BlurFade delay={BLUR_FADE_DELAY * 11}>
+            <div className="flex flex-col items-center justify-center space-y-4 text-center">
+              <div className="space-y-2">
+                <div className="inline-block rounded-lg bg-foreground text-background px-3 py-1 text-sm">
+                  {t('projects.label')}
+                </div>
+                <h2 className="text-3xl font-bold tracking-tighter sm:text-5xl">
+                  {t('projects.title')}
+                </h2>
+                <p className="text-muted-foreground md:text-xl/relaxed lg:text-base/relaxed xl:text-xl/relaxed">
+                  {t('projects.description')}
+                </p>
+              </div>
+            </div>
+          </BlurFade>
+          <div className="grid grid-cols-1 gap-3 sm:grid-cols-2 max-w-[800px] mx-auto">
+            {projectsList.map((project, id) => (
+              <BlurFade
+                key={project.title}
+                delay={BLUR_FADE_DELAY * 12 + id * 0.05}
+              >
+                <ProjectCard
+                  href={project.href}
+                  key={project.title}
+                  title={project.title}
+                  description={project.description}
+                  dates={project.dates}
+                  tags={project.technologies}
+                  image={project.image}
+                  links={Object.entries(project.links).map(([type, href]) => ({
+                    icon: <LinkIcon className="size-4" />,
+                    type,
+                    href,
+                  }))}
+                />
+              </BlurFade>
+            ))}
+          </div>
+        </div>
+      </section>
+      <section id="contact">
+        <div className="grid items-center justify-center gap-4 px-4 text-center md:px-6 w-full py-12">
+          <BlurFade delay={BLUR_FADE_DELAY * 16}>
+            <div className="space-y-3">
+              <div className="inline-block rounded-lg bg-foreground text-background px-3 py-1 text-sm">
+                {t('contact.label')}
+              </div>
+              <h2 className="text-3xl font-bold tracking-tighter sm:text-5xl">
+                {t('contact.title')}
+              </h2>
+              <p className="mx-auto max-w-[600px] text-muted-foreground md:text-xl/relaxed lg:text-base/relaxed xl:text-xl/relaxed">
+                {t('contact.description')}{" "}
+                <Link
+                  href={t('contact.linkedinUrl')}
+                  className="text-blue-500 hover:underline"
+                  target="_blank"
+                >
+                  {t('contact.linkedinText')}
+                </Link>{" "}
+                {t('contact.response')}
+              </p>
+            </div>
+          </BlurFade>
+        </div>
+      </section>
+    </main>
+  );
 }
