--- src/app/i18n/locales/en-US/footer.json
@@ -1,3 +1,0 @@
-{
-    "languageSwitcher": "Switch from <1>{{lng}}</1> to: "
-}
