--- src/app/i18n/locales/en-US/translation.json
@@ -1,4 +1,0 @@
-{
-    "title": "Hi there!",
-    "to-second-page": "To second page"
-}
