--- src/app/i18n/locales/pt-BR/footer.json
@@ -1,3 +1,0 @@
-{
-    "languageSwitcher": "Wechseln von <1>{{lng}}</1> nach: "
-}
