--- src/app/i18n/locales/pt-BR/translation.json
@@ -1,4 +1,0 @@
-{
-    "title": "Hallo Leute!",
-    "to-second-page": "Zur zweiten Seite"
-}
