+++ src/app/i18n/settings.ts
@@ -1,5 +1,5 @@
 export const fallbackLng = 'en'
-export const languages = [fallbackLng, 'de']
+export const languages = [fallbackLng, 'pt']
 export const defaultNS = 'translation'
 export const cookieName = 'i18next'
