+++ src/components/navbar.tsx
@@ -1,3 +1,5 @@
+import { useTranslation } from "@/app/i18n";
+import { languages } from "@/app/i18n/settings";
 import { Dock, DockIcon } from "@/components/magicui/dock";
 import { ModeToggle } from "@/components/mode-toggle";
 import { buttonVariants } from "@/components/ui/button";
@@ -11,7 +13,8 @@ import { DATA_en } from "@/data/resume_en";
 import { cn } from "@/lib/utils";
 import Link from "next/link";
 
-export default function Navbar() {
+export default function Navbar({ lng }: { lng: string }) {
+  
   return (
     <div className="pointer-events-none fixed inset-x-0 bottom-0 z-30 mx-auto mb-4 flex origin-bottom h-full max-h-14">
       <div className="fixed bottom-0 inset-x-0 h-16 w-full bg-background to-transparent backdrop-blur-lg [-webkit-mask-image:linear-gradient(to_top,black,transparent)] dark:bg-background"></div>
@@ -67,6 +70,18 @@ export default function Navbar() {
             </TooltipTrigger>
             <TooltipContent>
               <p>Theme</p>
+            </TooltipContent>
+          </Tooltip>
+        </DockIcon>
+        <DockIcon>
+          <Tooltip>
+            <TooltipTrigger asChild>
+              <Link href={`/${lng === "pt" ? "en" : "pt"}`} className={cn(buttonVariants({ variant: "ghost", size: "icon" }), "size-12")}>
+                {lng === "pt" ? "🇧🇷" : "🇺🇸"}
+              </Link>
+            </TooltipTrigger>
+            <TooltipContent>
+              <p>Language</p>
             </TooltipContent>
           </Tooltip>
         </DockIcon>
